% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_address.R
\name{addresses}
\alias{addresses}
\title{addresses Function}
\format{\describe{
\item{longitude}{numeric; longitude}
\item{latitude}{numeric; latitude.}
\item{id}{factor; address id.}
\item{order}{numeric; order.}
\item{address}{factor; address name.}
}}
\usage{
addresses()
}
\description{
This function download information about address location
}
\details{
Details of usage
}
\examples{
addresses()
}
\keyword{keyword}
