% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{area Function}
\format{\describe{
\item{ID}{factor; ID of school area.}
\item{School_Nr}{factor; Number of school. }
\item{Shape}{factor; }
\item{PK}{factor; Shows type of school with number.}
\item{School_Name}{factor; Name of school area.}
\item{Longitude}{numeric; Longitude of parish site.}
\item{Latitude}{numeric; Latitude of parish site.}
}}
\usage{
area(coords = F)
}
\arguments{
\item{coords}{shows basic_data about schools area. When set TRUE shows coords of schools area.}
}
\description{
This function download data about basic school areas in Poznań.
}
\details{
Details of usage
}
\examples{

Area_basic <- area () 

Area_with_coord <- area (TRUE)
}
\keyword{keyword}
