% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_paths.R
\name{bike_paths}
\alias{bike_paths}
\title{bike_paths Function}
\format{\describe{
\item{ID}{factor; ID of all bike paths in Poznan.}
\item{Name}{factor; Name of bike paths.}
\item{Lenght}{factor; Lenght of bike paths}
\item{Description}{factor; Description of bike paths}
}}
\usage{
bike_paths(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about Bike paths in Poznań}
}
\description{
This function download data about Bike Paths in Poznan.
}
\details{
Details of usage
}
\examples{
Bike_Paths <- bike_paths(Coord = F)
Area_Electoral_coords <- bike_paths(Coord = T)
}
\keyword{keyword}
