% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bike_place.R
\name{bike_station}
\alias{bike_station}
\title{bike_station Function}
\format{\describe{
\item{id}{factor; station id.}
\item{longitude}{numericr; longitude.}
\item{latitude}{numeric; latitude.}
\item{name}{factor; station name.}
\item{bike_racks}{factor; bike racks.}
\item{free_racks}{factor; free racks.}
\item{bike_types}{factor; bike types.}
\item{bikes}{factor; number of bikes.}
\item{update}{factor; last update.}
}}
\usage{
bike_station()
}
\description{
This function download statistic about bike station from NextBike
}
\details{
Details of usage
}
\examples{
bike_station()
}
\keyword{keyword}
