% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_stations.R
\name{bike_stations}
\alias{bike_stations}
\title{bike_stations Function}
\format{\describe{
\item{id}{factor; station id.}
\item{Longitude}{numericr; Longitude.}
\item{Latitude}{numeric; Latitude.}
\item{name}{factor; station name.}
\item{bike_racks}{factor; bike racks.}
\item{free_racks}{factor; free racks.}
\item{bike_types}{factor; bike types.}
\item{bikes}{factor; number of bikes.}
\item{update}{factor; last update.}
}}
\usage{
bike_stations()
}
\description{
This function download statistic about bike station from NextBike
}
\details{
Details of usage
}
\examples{
bike_stations()
}
\keyword{keyword}
