% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikes_current.R
\name{bikes_current}
\alias{bikes_current}
\title{bikes_current Function}
\format{\describe{
\item{number}{character; bike number.}
\item{bike_type}{character; bike type.}
\item{lock_types}{character; lock types.}
\item{id}{factor; station id.}
\item{uid}{factor; uid.}
\item{lat}{factor; latitude.}
\item{lng}{factor; longitude}
\item{name}{factor; name.}
\item{bikes}{factor; bikes.}
\item{bike_racks}{factor; bike racks.}
\item{free_racks}{factor; free racks.}
\item{bike_types}{factor; bike types.}
}}
\usage{
bikes_current()
}
\description{
This function download statistic about bikes from NextBike
}
\details{
Details of usage
}
\examples{
bikes_current()
}
\keyword{keyword}
