% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikes_data.R
\docType{data}
\name{bikes_data}
\alias{bikes_data}
\title{bikes_data Function}
\format{\describe{
\item{number}{character; bike number.}
\item{bike_type}{character; bike type.}
\item{lock_types}{character; lock types.}
\item{id}{factor; station id.}
\item{uid}{factor; uid.}
\item{lat}{factor; latitude.}
\item{lng}{factor; longitude.}
\item{name}{factor; name.}
\item{bikes}{factor; bikes.}
\item{bike_racks}{factor; bike racks.}
\item{free_racks}{factor; free racks.}
\item{bike_types}{factor; bike types.}
\item{date}{factor; date.}
}}
\usage{
bikes_data
}
\description{
This function load data about bikes.
}
\examples{
bikes_data()
}
\keyword{datasets}
