% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesspools.R
\name{cesspools}
\alias{cesspools}
\title{cesspools Function}
\format{\describe{
\item{ID}{factor; ID of cesspool.}
\item{Cesspool_Nr}{factor; Number of cesspool.}
\item{Cesspool_Street}{factor; Street of cesspool.}
\item{Cesspoll_Plot}{factor; Number of plot. }
\item{Longitude}{numeric; Longitude of cesspool.}
\item{Latitude}{numeric; Latitude of cesspool.}
}}
\usage{
cesspools()
}
\description{
This function download data about cesspools in Poznań.
}
\examples{
Cesspools <- cesspools()
}
\keyword{keyword}
