% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churches_data.R
\docType{data}
\name{churches_data}
\alias{churches_data}
\title{churches_data  Function}
\format{\describe{
\item{Parish_Name.y}{character; Name of certain Church.}
\item{Longitude}{numeric; Longitude of church.}
\item{Latitude}{numeric; Latitude of church.}
}}
\usage{
churches_data
}
\description{
This function load data about churches in Poznan.
}
\examples{
churches_data()
}
\keyword{datasets}
