% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churches_wiki.R
\name{churches_wiki}
\alias{churches_wiki}
\title{churches_wiki  Function}
\format{\describe{
\item{Parish_Name.y}{character; Name of certain Church.}
\item{Longitude}{numeric; Longitude of church.}
\item{Latitude}{numeric; Latitude of church.}
}}
\usage{
churches_wiki()
}
\description{
This function download data about churches in Poznan.
}
\examples{
Churches <- churches_wiki()
}
\keyword{keyword}
