% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districts.R
\name{districts}
\alias{districts}
\title{districts  Function}
\format{\describe{
\item{ID}{factor; ID of District.}
\item{Name}{factor; Name of District.}
\item{Longitude}{numeric; Longitude of District. site.}
\item{Latitude}{numeric; Latitude of District. site.}
}}
\usage{
districts(coords = F)
}
\arguments{
\item{coords}{show basic data about districts in Poznań. When set TRUE shows coords of District area.}
}
\description{
This function download data about districts  in Poznan.
}
\examples{
District <- districts(coords = F)
District_coord <- districts(coords = T)
coords <- District_coord$coord
}
\keyword{keyword}
