% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dydactic_paths.R
\name{dydactic_paths}
\alias{dydactic_paths}
\title{dydactic_paths  Function}
\format{\describe{
\item{ID}{factor; ID of Dydactic paths.}
\item{Name}{factor; Name of Dydactic paths in Poznan.}
\item{Length}{factor; Length of Dydactic paths.}
\item{Description}{factor; Description of Dydactic paths in Poznan.}
}}
\usage{
dydactic_paths(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about Dydactic pathsin Poznań}
}
\description{
This function download data about Dydactic paths  in Poznan.
}
\details{
Details of usage
}
\examples{
Dydactic_Paths <- dydactic_paths(Coord = F)
Dydactic_Paths_coord <- dydactic_paths(Coord = T)
}
\keyword{keyword}
