% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoral_areas.R
\name{electoral_areas}
\alias{electoral_areas}
\title{electoral_areas Function}
\format{\describe{
\item{ID}{numeric; ID of area electoral.}
\item{District_no}{numeric; Number of district.}
\item{Longitude}{numeric; Longitude of area electoral.}
\item{Latitude}{numeric; Latitude of area electoral.}
}}
\usage{
electoral_areas(coords = F)
}
\arguments{
\item{coords}{show basic data about electoral areas in Poznań.  When set TRUE shows coords of electoral areas.}
}
\description{
This function download data about electoral areas in Poznan.
}
\details{
Details of usage
}
\examples{
Area_Electoral <- electoral_areas(coords = F)
Areas <- electoral_areas(coords = T)
coord <- Areas$coord
}
\keyword{keyword}
