% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_address.R
\name{get_address}
\alias{get_address}
\title{get_address Function}
\format{\describe{
\item{id}{factor; address id.}
\item{longitude}{factor; longitude}
\item{latitude}{factor; latitude.}
\item{address}{factor; address name.}
}}
\usage{
get_address()
}
\description{
This function download information about address location
}
\details{
Details of usage
}
\examples{
get_address()
}
\keyword{keyword}
