% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lsdp.R
\name{get_lsdp}
\alias{get_lsdp}
\title{get_lsdp Function}
\format{\describe{
\item{id}{factor; id.}
\item{lsdp_location_name}{factor; location name.}
\item{lsdp_adaptive_resolution}{factor; adaptive resolution.}
\item{lsdp_date_of_ar}{factor; date of adaptive resolution.}
\item{lsdp_approval_of_the_resolution}{factor; approval of the resolution.}
\item{lsdp_date_of_the_ar}{factor; date of the approval of the resolution.}
\item{lsdp_official_journal}{factor; official journal.}
\item{lsdp_date_of_oj}{factor; date of official journal.}
}}
\usage{
get_lsdp(basic = TRUE)
}
\description{
This function download data about Passed Local Spatial Development Plans
}
\details{
Details of usage
}
\examples{
get_lsdp()
}
\keyword{keyword}
