% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lsdpc.R
\name{get_lsdpc}
\alias{get_lsdpc}
\title{get_lsdpc Function}
\format{\describe{
\item{id}{factor; id.}
\item{lsdpc_location_name}{factor; location name.}
\item{lsdpc_adopted_change}{factor; adopted change.}
\item{lsdpc_adaptive_resolution}{factor; adaptive resolution.}
\item{lsdpc_date_oAR}{factor; date of adaptive resolution.}
}}
\usage{
get_lsdpc(coords = F)
}
\description{
This function download data about Called Local Spatial Development Plans
}
\details{
Details of usage
}
\examples{
get_lsdpc()
}
\keyword{keyword}
