% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bike_place_details_function.R
\name{get_nextbike_data}
\alias{get_nextbike_data}
\title{get_nextbike_data Function}
\format{\describe{
\item{number}{character; bike number.}
\item{bike_type}{character; bike type.}
\item{lock_types}{character; lock types.}
\item{id}{factor; station id.}
\item{uid}{factor; uid.}
\item{lat}{factor; latitude.}
\item{lng}{factor; longitude.}
\item{name}{factor; name.}
\item{bikes}{factor; bikes.}
\item{bike_racks}{factor; bike racks.}
\item{free_racks}{factor; free racks.}
\item{bike_types}{factor; bike types.}
\item{date}{factor; date.}
}}
\usage{
get_nextbike_data(start, end)
}
\arguments{
\item{start}{the beginning of data downloading}

\item{end}{the end of data downloading}
}
\description{
This function download statistic about bikes from NextBike
}
\details{
Details of usage
}
\examples{
get_nextbike_data("27-07-2018 13:05","28-07-2018 13:05")
}
\keyword{keyword}
