% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_property_cto.R
\name{get_property_cto}
\alias{get_property_cto}
\title{get_property_cto Function}
\format{\describe{
\item{id}{factor; property id.}
\item{longitude}{factor; longitude.}
\item{latitude}{factor; latitude.}
\item{plot_number}{factor; plot_number.}
\item{plot_of_land}{factor; plot_of_land.}
\item{description}{factor; description.}
\item{comment}{factor; comments.}
\item{url_bip}{factor; url bip.}
}}
\usage{
get_property_cto()
}
\description{
This function download data about property current trade offer
}
\details{
Details of usage
}
\examples{
get_property_cto()
}
\keyword{keyword}
