% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sdbs.R
\name{get_sdbs}
\alias{get_sdbs}
\title{get_sdbs Function}
\format{\describe{
\item{station}{factor; station name.}
\item{longitude}{factor; longitude}
\item{latitude}{factor; latitude.}
\item{station}{factor; station name.}
\item{longitude}{factor; longitude}
\item{latitude}{factor; latitude.}
\item{straight_distance}{factor; the shortest straight distance between stations.}
}}
\usage{
get_sdbs()
}
\description{
This function download data from NextBike and and calculates the shortest distance between stations
}
\details{
Details of usage
}
\examples{
get_sdbs
}
\keyword{keyword}
