% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArea_electoral.R
\name{getArea_electoral}
\alias{getArea_electoral}
\title{Area Electoral Function}
\format{\describe{
\item{ID}{factor; ID of area electoral.}
\item{District_no}{factor; Number of District.}
}}
\usage{
getArea_electoral(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about Area electoral in Poznań}
}
\description{
This function download data about Area Electoral in Poznan.
}
\details{
Details of usage
}
\examples{
Area_Electoral <- getArea_electoral(Coord = F)
Area_Electoral_coords <- getArea_electoral(Coord = T)
}
\keyword{keyword}
