% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBikePaths.R
\name{getBike_Paths}
\alias{getBike_Paths}
\title{Bike Paths Function}
\format{\describe{
\item{ID}{factor; ID of all bike paths in Poznan.}
\item{Name}{factor; Name of bike paths.}
\item{Lenght}{factor; Lenght of bike paths}
\item{Description}{factor; Description of bike paths}
}}
\usage{
getBike_Paths(Coord = F)
}
\description{
This function download data about Bike Paths in Poznan.
}
\details{
Details of usage
}
\examples{
Bike_Paths <- getBike_Paths(Coord = F)
Area_Electoral_coords <- getBike_Paths(Coord = T)
}
\keyword{keyword}
