% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChurch.R
\name{getChurch}
\alias{getChurch}
\title{Church  Function}
\format{\describe{
\item{ID}{factor; ID of church.}
\item{Code}{factor; code of church in Poznan.}
\item{Name}{factor; Name of church.}
\item{Category_builiding}{factor; Category of church in Poznan.}
\item{City}{factor; City.}
\item{Lang}{factor; Language.}
\item{Address}{factor; Address of church.}
\item{ID_monument}{factor; ID monument in Poznan.}
\item{Description}{factor; Description of church.}
}}
\usage{
getChurch(Coord = F)
}
\description{
This function download data about Church in Poznan.
}
\details{
Details of usage
}
\examples{
Church <- getChurch(Coord = F)
Church_coord <- getChurch(Coord = T)
}
\keyword{keyword}
