% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircle.R
\name{getCircle}
\alias{getCircle}
\title{Circle electoral  Function}
\format{\describe{
\item{ID}{factor; ID of circle.}
\item{Name}{factor; Name of circle.}
}}
\usage{
getCircle(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about circle electoral of local government in Poznań}
}
\description{
This function download data about Circle electoral in Poznan.
}
\details{
Details of usage
}
\examples{
Circle_electoral <- getCircle(Coord = F)
Circle_coord <- getCircle(Coord = T)
}
\keyword{keyword}
