% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCouncil.R
\name{getCouncil}
\alias{getCouncil}
\title{Council  Function}
\format{\describe{
\item{ID}{factor; ID of  Estate.}
\item{Type}{factor; Type of Estate in Poznan.}
\item{Address}{factor; Address of Residence.}
\item{Residence}{factor; Residence of Estate in Poznan.}
\item{Electoral_District}{factor; Electoral District in Poznan.}
\item{AFDP}{factor; Adaptation for disabled pearson.}
\item{Venue_no}{factor; Venue of Estate in Poznan.}
\item{Distroct_no}{factor; District number in Poznan.}
}}
\usage{
getCouncil(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about district council in Poznań}
}
\description{
This function download data about district council in Poznan.
}
\details{
Details of usage
}
\examples{
Estate <- getEstate(Coord = F)
Estate_coord <- getEstate(Coord = T)
}
\keyword{keyword}
