% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistrict.R
\name{getDistrict}
\alias{getDistrict}
\title{District  Function}
\format{\describe{
\item{ID}{factor; ID of District.}
\item{Name}{factor; Name of District.}
\item{Type}{factor; Type of District.}
}}
\usage{
getDistrict(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about distric in Poznań}
}
\description{
This function download data about District  in Poznan.
}
\details{
Details of usage
}
\examples{
District <- getDistrict(Coord = F)
District_coord <- getDistrict(Coord = T)
}
\keyword{keyword}
