% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDydactic_Paths.R
\name{getDydactic_Paths}
\alias{getDydactic_Paths}
\title{Dydactic Paths  Function}
\format{\describe{
\item{ID}{factor; ID of Dydactic paths.}
\item{Name}{factor; Name of Dydactic paths in Poznan.}
\item{Length}{factor; Length of Dydactic paths.}
\item{Description}{factor; Description of Dydactic paths in Poznan.}
}}
\usage{
getDydactic_Paths(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about Dydactic pathsin Poznań}
}
\description{
This function download data about Dydactic paths  in Poznan.
}
\details{
Details of usage
}
\examples{
Dydactic_Paths <- getDydactic_Paths(Coord = F)
Dydactic_Paths_coord <- getDydactic_Paths(Coord = T)
}
\keyword{keyword}
