% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOblast.R
\name{getOblast}
\alias{getOblast}
\title{Oblast Function}
\format{\describe{
\item{ID}{factor; ID of Oblast.}
\item{Name}{factor; Name of Oblast in Poznan.}
}}
\usage{
getOblast(Coord = F)
}
\description{
This function download data about Oblast in Poznan.
}
\details{
Details of usage
}
\examples{
oblast <- getOblast(Coord = F)
oblast_coord <- getOblast(Coord = T)
}
\keyword{keyword}
