% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPaths.R
\name{getPaths}
\alias{getPaths}
\title{Paths  Function}
\usage{
getPaths(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about trail in Poznań}
}
\description{
This function download data about trail in Poznan.
}
\details{
Details of usage
}
\examples{
Paths <- getPaths(Coord = F)
Paths_coord <- getPaths(Coord = T)
}
\keyword{keyword}
