% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelict.R
\name{getRelict}
\alias{getRelict}
\title{Relict  Function}
\format{\describe{
\item{ID}{factor; ID of relikt.}
\item{Code}{factor; code of relikt. in Poznan.}
\item{Name}{factor; Name of relikt.}
\item{Category_builiding}{factor; Category of relikt. in Poznan.}
\item{City}{factor; City.}
\item{Lang}{factor; Language.}
\item{Address}{factor; Address of relikt.}
\item{ID_monument}{factor; ID monument in Poznan.}
\item{Description}{factor; Description of relikt.}
}}
\usage{
getRelict(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about Relict in Poznań}
}
\description{
This function download data about Relict in Poznan.
}
\details{
Details of usage
}
\examples{
relikt. <- getRelict(Coord = F)
relikt._coord <- getRelict(Coord = T)
}
\keyword{keyword}
