% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTicket.R
\name{getTicket}
\alias{getTicket}
\title{Ticket  Function}
\format{\describe{
\item{Opening_Hours}{factor; Hours of open.}
\item{Sales_form}{factor; form of sale.}
\item{Ticket}{factor; kind of Ticket.}
\item{Name}{factor; Name.}
\item{Address}{factor; Address of shpo.}
\item{City}{factor;City.}
\item{Lang}{factor; Language.}
\item{PEKKA_Card}{factor; Card.}
\item{ID_Class}{factor; ID of Class.}
\item{Description}{factor; Decsription.}
\item{POS}{factor; Description Class.}
}}
\usage{
getTicket(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about Ticket in Poznań}
}
\description{
This function download data about Ticket in Poznan.
}
\details{
Details of usage
}
\examples{
Ticket <- getTicket(Coord = F)
Ticket_coord <- getTicket(Coord = T)
}
\keyword{keyword}
