% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.R
\name{getTrakt}
\alias{getTrakt}
\title{Royal trakt Function}
\format{\describe{
}}
\usage{
getTrakt()
}
\description{
This function download data about Royal trakt in Poznan.
}
\details{
Details of usage
}
\examples{
Trakt <- getTrakt()
}
\keyword{keyword}
