% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graves.R
\name{graves}
\alias{graves}
\title{graves  Function}
\format{\describe{
\item{Cemetery_ID}{factor; ID of  Cemetery in Poznan.}
\item{Cemetery_No}{factor; Cemetery number.}
\item{Date_birth}{factor; Date birth pearson.}
\item{Date_death}{factor; Date death pearson.}
\item{Date_Burial}{factor; Date burial pearson.}
\item{Row_No}{factor; Row.}
\item{Grave_Name}{factor; Name.}
\item{Grave_Surname}{factor; Surname.}
\item{Grave_Name_Surname}{factor; Name and Surname.}
\item{Field_No}{factor; Field number in cemetery.}
\item{Place_No}{factor; Place number in cemetery.}
\item{Paid}{factor; Paid.}
}}
\usage{
graves(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about graves in Poznań}
}
\description{
This function download data about graves in Poznan.
}
\details{
Details of usage
}
\examples{
Grave <- graves(Coord = F)
Grave_coord <- graves(Coord = T)
}
\keyword{keyword}
