% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graves_all.R
\name{graves_all}
\alias{graves_all}
\title{Graves All  Function}
\format{\describe{
\item{Cemetery_ID}{factor; ID of  Cemetery in Poznan.}
\item{Cemetery_No}{factor; Cemetery number.}
\item{Date_birth}{factor; Date birth pearson.}
\item{Date_death}{factor; Date death pearson.}
\item{Date_Burial}{factor; Date burial pearson.}
\item{Row_No}{factor; Row.}
\item{Grave_Name}{factor; Name.}
\item{Grave_Surname}{factor; Surname.}
\item{Grave_Name_Surname}{factor; Name and Surname.}
\item{Field_No}{factor; Field number in cemetery.}
\item{Place_No}{factor; Place number in cemetery.}
\item{Paid}{factor; Paid.}
}}
\usage{
graves_all(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about Graves All in Poznań}
}
\description{
This function download data about Graves All in Poznan.
}
\details{
Details of usage
}
\examples{
Grave_All <- getGrave_All(Coord = F)
Grave_All_coord <- getGrave_All(Coord = T)
}
\keyword{keyword}
