% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have_ip.R
\name{have_ip}
\alias{have_ip}
\title{have_ip Function}
\format{\describe{
}}
\usage{
have_ip()
}
\description{
This function check connection to internet.
}
\details{
Details of usage
}
\examples{
have_ip()
}
\keyword{keyword}
