% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/havingIP.R
\name{havingIP}
\alias{havingIP}
\title{havingIP Function}
\format{\describe{
}}
\usage{
havingIP()
}
\description{
This function check connection to internet.
}
\details{
Details of usage
}
\examples{
havingIP ()
}
\keyword{keyword}
