% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_churches.R
\name{historical_churches}
\alias{historical_churches}
\title{historical_churches  Function}
\format{\describe{
\item{ID}{numeric; ID of historical churches .}
\item{Code}{factor; code of historical church in Poznan.}
\item{Name}{factor; Name of historical church.}
\item{Category_builiding}{factor; Category of historical church in Poznan.}
\item{City}{factor; City.}
\item{Lang}{factor; Language.}
\item{Address}{factor; Address of historical church.}
\item{ID_monument}{factor; ID monument in Poznan.}
\item{Description}{factor; Description of historical church.}
\item{Longitude}{numeric; Longitude of historical church.}
\item{Latitude}{numeric; Latitude of historical church.}
}}
\usage{
historical_churches()
}
\description{
This function download data about historical Churches in Poznan.
}
\examples{
Church <- historical_churches()
Church_coord <- historical_churches()
}
\keyword{keyword}
