% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTrue.R
\name{isTrue}
\alias{isTrue}
\title{isTrue Function}
\format{\describe{
\item{type}{factor; type of.}
\item{long}{numeric; longitude.}
\item{lat}{numeric; latitude.}
}}
\usage{
isTrue(arg = TRUE)
}
\arguments{
\item{arg}{Is value TRUE? Defaults to TRUE.}
}
\description{
This function checks whether TRUE is TRUE.
}
\details{
Details of usage
}
\examples{
isTrue()
}
\keyword{keyword}
