% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_government.R
\name{local_government}
\alias{local_government}
\title{Local Government Function}
\format{\describe{
\item{ID}{factor; ID of District.}
\item{AFDP}{factor; Adaptation for disabled pearson.}
\item{District_no}{factor; Number of district.}
\item{Residence}{factor; Name of Residence. }
}}
\usage{
local_government(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about local government in Poznań}
}
\description{
This function download data about local government in Poznań.
}
\details{
Details of usage
}
\examples{
Local_Government <- local_government(Coord = F)
}
\keyword{keyword}
