% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_spatial_dev_plans_called.R
\name{local_spatial_dev_plans_called}
\alias{local_spatial_dev_plans_called}
\title{local_spatial_dev_plans_called Function}
\format{\describe{
\item{id}{factor; id.}
\item{lsdpc_location_name}{factor; location name.}
\item{lsdpc_adopted_change}{factor; adopted change.}
\item{lsdpc_adaptive_resolution}{factor; adaptive resolution.}
\item{lsdpc_date_oAR}{factor; date of adaptive resolution.}
}}
\usage{
local_spatial_dev_plans_called(basic = TRUE)
}
\description{
This function download data about Called Local Spatial Development Plans
}
\details{
Details of usage
}
\examples{
local_spatial_dev_plans_called()
}
\keyword{keyword}
