% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monuments.R
\name{monuments}
\alias{monuments}
\title{monuments  Function}
\format{\describe{
\item{ID}{factor; ID of relikt.}
\item{Code}{factor; code of relikt. in Poznan.}
\item{Name}{factor; Name of relikt.}
\item{Category_builiding}{factor; Category of relikt. in Poznan.}
\item{City}{factor; City.}
\item{Lang}{factor; Language.}
\item{Address}{factor; Address of relikt.}
\item{ID_monument}{factor; ID monument in Poznan.}
\item{Description}{factor; Description of relikt.}
}}
\usage{
monuments(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about monuments in Poznań}
}
\description{
This function download data about monuments in Poznan.
}
\details{
Details of usage
}
\examples{
relikt. <- monuments(Coord = F)
relikt._coord <- monuments(Coord = T)
}
\keyword{keyword}
