% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parish.R
\name{parish}
\alias{parish}
\title{parish Function}
\format{\describe{
\item{ID}{character; ID of parish.}
\item{Parish_Name}{factor; Number of sewage work.}
\item{Longitude}{numeric; Longitude of parish site.}
\item{Latitude}{numeric; Latitude of parish site.}
}}
\usage{
parish(coords = F)
}
\arguments{
\item{coords}{shows basic_data about parishes. When set TRUE shows coords of Parishes}
}
\description{
This function download data about parishes in Poznań
}
\details{
Details of usage
}
\examples{

parish_basic <- parish()

parish_with_coord <- parish(TRUE)

}
\keyword{keyword}
