% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parishes.R
\name{parishes}
\alias{parishes}
\title{parishes Function}
\format{\describe{
\item{ID}{character; ID of parish.}
\item{Parish_Name}{factor; Number of sewage work.}
\item{Longitude}{numeric; Longitude of parish site.}
\item{Latitude}{numeric; Latitude of parish site.}
}}
\usage{
parishes(coords = F)
}
\arguments{
\item{coords}{shows basic_data about parishes. When set TRUE shows coords of Parishes}
}
\description{
This function download data about parishes in Poznań
}
\details{
Details of usage
}
\examples{

parish_basic <- parishes()

parish_with_coord <- parishes(TRUE)

}
\keyword{keyword}
