% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parking_machines.R
\name{parking_machines}
\alias{parking_machines}
\title{parking_machines Function}
\format{\describe{
\item{ID}{factor; ID of PM.}
\item{PM_Zone}{factor; Zone in which PM is placed.}
\item{PM_Street}{factor; Street in which PM is placed.}
\item{Longitude}{numeric; Longitude of PM.}
\item{Latitude}{numeric; Latitude of PM.}
}}
\usage{
parking_machines()
}
\description{
This function download data about parking machines in Poznań
}
\examples{

pms <- parking_machines()



}
\keyword{keyword}
