% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_dydactic.R
\name{paths_dydactic}
\alias{paths_dydactic}
\title{paths_dydactic  Function}
\format{\describe{
\item{ID}{numeric; ID of dydactic paths.}
\item{Name}{factor; Name of dydactic paths in Poznan.}
\item{Length}{factor; Length of dydactic paths.}
\item{Description}{factor; Description of dydactic paths in Poznan.}
}}
\usage{
paths_dydactic(coords = F)
}
\arguments{
\item{coords}{show basic data about dydactic pathsin Poznań}
}
\description{
This function download data about dydactic paths  in Poznan.
}
\details{
Details of usage
}
\examples{
Dydactic_Paths <- paths_dydactic(coords = F)
Dydactic_Paths_coord <- paths_dydactic(coords = T)
}
\keyword{keyword}
