% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_walking.R
\name{paths_walking}
\alias{paths_walking}
\title{paths_walking  Function}
\usage{
paths_walking(coords = F)
}
\arguments{
\item{coords}{show basic data about paths walking in Poznań}
}
\description{
This function download data about paths walking in Poznan.
}
\details{
Details of usage
}
\examples{
Paths <- paths_walking(coords = F)
Paths_coord <- paths_walking(coords = T)
}
\keyword{keyword}
