% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{pm}
\alias{pm}
\title{pm Function}
\format{\describe{
\item{ID}{factor; ID of PM.}
\item{PM_Zone}{factor; Zone in which PM is placed.}
\item{PM_Street}{factor; Street in which PM is placed.}
\item{Longitude}{numeric; Longitude of PM.}
\item{Latitude}{numeric; Latitude of PM.}
}}
\usage{
pm()
}
\description{
This function download data about parking machines in Poznań
}
\details{
Details of usage
}
\examples{

pms <- pm()



}
\keyword{keyword}
