% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties_current_trade.R
\name{properties_current_trade}
\alias{properties_current_trade}
\title{properties_current_trade Function}
\format{\describe{
\item{id}{factor; property id.}
\item{longitude}{numeric; longitude.}
\item{latitude}{numeric; latitude.}
\item{plot_number}{factor; plot_number.}
\item{plot_of_land}{factor; plot_of_land.}
\item{description}{factor; description.}
\item{comment}{factor; comments.}
\item{url_bip}{factor; url bip.}
}}
\usage{
properties_current_trade()
}
\description{
This function download data about property current trade offer
}
\details{
Details of usage
}
\examples{
properties_current_trade()
}
\keyword{keyword}
