% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_property_fs.R
\name{properties_for_sale}
\alias{properties_for_sale}
\title{properties_for_sale Function}
\format{\describe{
\item{id}{factor; property id.}
\item{longitude}{numeric; longitude.}
\item{latitude}{numeric; latitude.}
\item{plot_number}{factor; plot_number.}
\item{plot_of_land}{factor; plot_of_land.}
\item{description}{factor; description.}
\item{comment}{factor; comments.}
\item{url_bip}{factor; url bip.}
}}
\usage{
properties_for_sale()
}
\description{
This function download data about property for sale
}
\details{
Details of usage
}
\examples{
properties_for_sale()
}
\keyword{keyword}
