% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royal_imperial_route.R
\name{royal_imperial_route}
\alias{royal_imperial_route}
\title{royal_imperial_route Function}
\format{\describe{
}}
\usage{
royal_imperial_route()
}
\description{
This function download data about Royal trakt in Poznan.
}
\details{
Details of usage
}
\examples{
Trakt <- royal_imperial_route()
}
\keyword{keyword}
