% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_basic_areas.R
\name{school_basic_areas}
\alias{school_basic_areas}
\title{school_basic_areas Function}
\format{\describe{
\item{ID}{factor; ID of school area.}
\item{School_Nr}{factor; Number of school. }
\item{Shape}{factor; }
\item{PK}{factor; Shows type of school with number.}
\item{School_Name}{factor; Name of school area.}
\item{Longitude}{numeric; Longitude of school area. Available only if coords is set TRUE.}
\item{Latitude}{numeric; Latitude of school area. Available only if coords is set TRUE.}
}}
\usage{
school_basic_areas(coords = F)
}
\arguments{
\item{coords}{shows basic_data about schools area. When set TRUE shows coords of schools area.}
}
\description{
This function download data about basic school areas in Poznań.
}
\examples{

Area_basic <- school_basic_areas () 

Area_with_coord <- school_basic_areas (TRUE)
}
\keyword{keyword}
