% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sewage_works.R
\name{sewage_works}
\alias{sewage_works}
\title{sewage_works Function}
\format{\describe{
\item{ID}{factor; ID of sewage work.}
\item{Cesspool_Nr}{factor; Number of sewage work.}
\item{Cesspool_Street}{factor; Street of sewage work.}
\item{Cesspoll_Plot}{factor; Number of plot. }
\item{Longitude}{numeric; Longitude of sewage work.}
\item{Latitude}{numeric; Latitude of sewage work.}
}}
\usage{
sewage_works()
}
\description{
This function download data about sewage works in Poznań.
}
\examples{

sws <- sewage_works()



}
\keyword{keyword}
