% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop.R
\name{stop}
\alias{stop}
\title{stop Function}
\format{\describe{
\item{ID}{factor; ID of stop.}
\item{Stop_Zone}{factor; Zone in which stop is placed.}
\item{Route_Type}{factor; Shows route type on which stop is placed.}
\item{Stop_Headsign}{factor; Shows which trams and buses departure from stop.}
\item{Stop_Name}{factor; Name of stop.}
\item{Longitude}{numeric; Longitude of stop.}
\item{Latitude}{numeric; Latitude of stop.}
}}
\usage{
stop()
}
\description{
This function download data about stops in Poznań
}
\details{
Details of usage
}
\examples{

Stops <- stop() 


}
\keyword{keyword}
