% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw.R
\name{sw}
\alias{sw}
\title{sw Function}
\format{\describe{
\item{ID}{factor; ID of sewage work.}
\item{Cesspool_Nr}{factor; Number of sewage work.}
\item{Cesspool_Street}{factor; Street of sewage work.}
\item{Cesspoll_Plot}{factor; Number of plot. }
\item{Longitude}{numeric; Longitude of sewage work.}
\item{Latitude}{numeric; Latitude of sewage work.}
}}
\usage{
sw()
}
\description{
This function download data about sewage works in Poznań.
}
\details{
Details of usage
}
\examples{

sws <- sw()



}
\keyword{keyword}
