% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticket_sales_points.R
\name{ticket_sales_points}
\alias{ticket_sales_points}
\title{ticket_sales_points  Function}
\format{\describe{
\item{Opening_Hours}{factor; Hours of open.}
\item{Sales_form}{factor; form of sale.}
\item{Ticket}{factor; kind of Ticket.}
\item{Name}{factor; Name.}
\item{Address}{factor; Address of shpo.}
\item{City}{factor;City.}
\item{Lang}{factor; Language.}
\item{PEKKA_Card}{factor; Card.}
\item{ID_Class}{factor; ID of Class.}
\item{Description}{factor; Decsription.}
\item{Longitude}{numeric; Longitude of sales points.}
\item{Latitude}{numeric; Latitude of sales points.}
}}
\usage{
ticket_sales_points()
}
\description{
This function download data about ticket sales points in Poznan.
}
\examples{
Ticket <- ticket_sales_points()
}
\keyword{keyword}
