% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticket_vending_machines.R
\name{ticket_vending_machines}
\alias{ticket_vending_machines}
\title{ticket_vending_machines Function}
\format{\describe{
\item{ID}{factor; ID of TVM.}
\item{TVM}{factor; Name of TVM.}
\item{PEKA}{factor; Shows if TVM accepts PEKA card.}
\item{Store_Card}{factor; Shows if TVM accepts store card.}
\item{TVM_Order}{factor; Order of TVM.}
\item{TVM_Description }{factor; Description where you can find this TVM..}
\item{Longitude}{numeric; Longitude of TVM.}
\item{Latitude}{numeric; Latitude of TVM.}
}}
\usage{
ticket_vending_machines()
}
\description{
This function download data about ticket vending machines in Poznań
}
\details{
Details of usage
}
\examples{

tvms <- ticket_vending_machines()



}
\keyword{keyword}
