% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemeteries.R
\name{cemeteries}
\alias{cemeteries}
\title{cemeteries Function}
\format{\describe{
\item{ID}{factor; ID of Cemetery in Poznan.}
\item{Cemetery_Name}{factor; Name of cemetery in Poznan.}
\item{Cemetery_Type}{factor; Type of cemetery in Poznan.}
}}
\usage{
cemeteries(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about cemetery in Poznań}
}
\description{
This function download data about Cemetery in Poznan.
}
\details{
Details of usage
}
\examples{
Cemetery <- getCemetery(Coord = F)
Cemetery_coords <- getCemetery(Coord = T)
}
\keyword{keyword}
