% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCesspool.R
\name{cesspool}
\alias{cesspool}
\title{cesspool Function}
\format{\describe{
\item{ID}{factor; ID of cesspool.}
\item{Cesspool_Nr}{factor; Number of cesspool.}
\item{Cesspool_Street}{factor; Street of cesspool.}
\item{Cesspoll_Plot}{factor; Number of plot. }
\item{Longitude}{numeric; Longitude of cesspool.}
\item{Latitude}{numeric; Latitude of cesspool.}
}}
\usage{
cesspool()
}
\description{
This function download data about cesspools in Poznań.
}
\details{
Details of usage
}
\examples{
cesspool()
}
\keyword{keyword}
