% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/council_districts.R
\name{council_districts}
\alias{council_districts}
\title{council_districts  Function}
\format{\describe{
\item{ID}{factor; ID of  Estate.}
\item{Type}{factor; Type of Estate in Poznan.}
\item{Address}{factor; Address of Residence.}
\item{Residence}{factor; Residence of Estate in Poznan.}
\item{Electoral_District}{factor; Electoral District in Poznan.}
\item{AFDP}{factor; Adaptation for disabled pearson.}
\item{Venue_no}{factor; Venue of Estate in Poznan.}
\item{Distroct_no}{factor; District number in Poznan.}
}}
\usage{
council_districts(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about council districts in Poznań}
}
\description{
This function download data about council districts in Poznan.
}
\details{
Details of usage
}
\examples{
Estate <- getEstate(Coord = F)
Estate_coord <- getEstate(Coord = T)
}
\keyword{keyword}
