% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districts.R
\name{districts}
\alias{districts}
\title{districts  Function}
\format{\describe{
\item{ID}{factor; ID of District.}
\item{Name}{factor; Name of District.}
\item{Type}{factor; Type of District.}
}}
\usage{
districts(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about districts in Poznań}
}
\description{
This function download data about districts  in Poznan.
}
\details{
Details of usage
}
\examples{
District <- getDistrict(Coord = F)
District_coord <- getDistrict(Coord = T)
}
\keyword{keyword}
