% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districts.R
\name{districts}
\alias{districts}
\title{districts  Function}
\format{\describe{
\item{ID}{factor; ID of District.}
\item{Name}{factor; Name of District.}
}}
\usage{
districts(coords = F)
}
\arguments{
\item{coords}{show basic data about districts in Poznań}
}
\description{
This function download data about districts  in Poznan.
}
\details{
Details of usage
}
\examples{
District <- districts(coords = F)
District_coord <- districts(coords = T)
}
\keyword{keyword}
