% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoral_areas.R
\name{electoral_areas}
\alias{electoral_areas}
\title{electoral_areas Function}
\format{\describe{
\item{ID}{factor; ID of area electoral.}
\item{District_no}{factor; Number of District.}
}}
\usage{
electoral_areas(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about Area electoral in Poznań}
}
\description{
This function download data about Area Electoral in Poznan.
}
\details{
Details of usage
}
\examples{
Area_Electoral <- electoral_areas(Coord = F)
Area_Electoral_coords <- electoral_areas(Coord = T)
}
\keyword{keyword}
