% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoral_areas.R
\name{electoral_areas}
\alias{electoral_areas}
\title{electoral_areas Function}
\format{\describe{
\item{ID}{numeric; ID of area electoral.}
\item{District_no}{numeric; Number of district.}
}}
\usage{
electoral_areas(coords = F)
}
\arguments{
\item{coords}{show basic data about electoral areas in Poznań}
}
\description{
This function download data about electoral areas in Poznan.
}
\details{
Details of usage
}
\examples{
Area_Electoral <- electoral_areas(coords = F)
Area_Electoral_coords <- electoral_areas(coords = T)
}
\keyword{keyword}
