% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoral_circles.R
\name{electoral_circles}
\alias{electoral_circles}
\title{electoral_circles  Function}
\format{\describe{
\item{ID}{factor; ID of circle.}
\item{Name}{factor; Name of circle.}
}}
\usage{
electoral_circles(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about electoral circles of local government in Poznań}
}
\description{
This function download data about electoral circlesl in Poznan.
}
\details{
Details of usage
}
\examples{
Circle_electoral <- electoral_circles(Coord = F)
Circle_coord <- electoral_circles(Coord = T)
}
\keyword{keyword}
