% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoral_oblasts.R
\name{electoral_oblasts}
\alias{electoral_oblasts}
\title{electoral_oblasts Function}
\format{\describe{
\item{ID}{numeric; ID of Oblast.}
\item{Name}{factor; Name of Oblast in Poznan.}
\item{Longitude}{numeric; Longitude of Oblast.}
\item{Latitude}{numeric; Latitude of Oblast.}
}}
\usage{
electoral_oblasts(coords = F)
}
\arguments{
\item{coords}{show basic data about electoral Oblast in Poznań. When set TRUE shows coords of schools area.}
}
\description{
This function download data about electoral Oblast in Poznan.
}
\examples{
oblast <- electoral_oblasts(coords = F)
oblast_coord <- electoral_oblasts(coords = T)
}
\keyword{keyword}
