% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graves.R
\name{graves}
\alias{graves}
\title{graves  Function}
\format{\describe{
\item{Cemetery_ID}{numeric; ID of  cemetery in Poznan.}
\item{Cemetery_No}{numeric; cemetery number.}
\item{Date_birth}{factor; Date birth pearson.}
\item{Date_death}{factor; Date death pearson.}
\item{Date_Burial}{factor; Date burial pearson.}
\item{Row_No}{factor; Row.}
\item{Grave_Name}{factor; Name.}
\item{Grave_Surname}{factor; Surname.}
\item{Grave_Name_Surname}{factor; Name and Surname.}
\item{Field_No}{factor; Field number in cemetery.}
\item{Place_No}{factor; Place number in cemetery.}
\item{Paid}{numeric; Paid.}
\item{Longitude}{numeric; Longitude of Grave.}
\item{Latitude}{numeric; Latitude of Grave.}
}}
\usage{
graves()
}
\description{
This function download data about graves in Poznan.
}
\examples{
Grave <- graves()
}
\keyword{keyword}
