% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graves_all.R
\name{graves_all}
\alias{graves_all}
\title{graves_all  Function}
\format{\describe{
\item{Cemetery_ID}{numeric; ID of  cemetery in Poznan.}
\item{Cemetery_No}{numeric; Cemetery number.}
\item{Date_birth}{factor; Date birth pearson.}
\item{Date_death}{factor; Date death pearson.}
\item{Date_Burial}{factor; Date burial pearson.}
\item{Row_No}{factor; Row.}
\item{Grave_Name}{factor; Name.}
\item{Grave_Surname}{factor; Surname.}
\item{Grave_Name_Surname}{factor; Name and Surname.}
\item{Field_No}{factor; Field number in cemetery.}
\item{Place_No}{factor; Place number in cemetery.}
\item{Paid}{numeric; Paid.}
}}
\usage{
graves_all(coords = F)
}
\arguments{
\item{coords}{show basic data about all graves in Poznań}
}
\description{
This function download data about all graves in Poznan.
}
\details{
Details of usage
}
\examples{
Grave_All <- graves_all(coords = F)
Grave_All_coord <- graves_all(coords = T)
}
\keyword{keyword}
