% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_churches.R
\name{historical_churches}
\alias{historical_churches}
\title{historical_churches  Function}
\format{\describe{
\item{ID}{factor; ID of church.}
\item{Code}{factor; code of church in Poznan.}
\item{Name}{factor; Name of church.}
\item{Category_builiding}{factor; Category of church in Poznan.}
\item{City}{factor; City.}
\item{Lang}{factor; Language.}
\item{Address}{factor; Address of church.}
\item{ID_monument}{factor; ID monument in Poznan.}
\item{Description}{factor; Description of church.}
}}
\usage{
historical_churches(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about historical churches in Poznań}
}
\description{
This function download data about historical Churches in Poznan.
}
\details{
Details of usage
}
\examples{
Church <- historical_churches(Coord = F)
Church_coord <- historical_churches(Coord = T)
}
\keyword{keyword}
