% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_churches.R
\name{historical_churches}
\alias{historical_churches}
\title{historical_churches  Function}
\format{\describe{
\item{ID}{numeric; ID of historical churches .}
\item{Code}{factor; code of historical church in Poznan.}
\item{Name}{factor; Name of historical church.}
\item{Category_builiding}{factor; Category of historical church in Poznan.}
\item{City}{factor; City.}
\item{Lang}{factor; Language.}
\item{Address}{factor; Address of historical church.}
\item{ID_monument}{factor; ID monument in Poznan.}
\item{Description}{factor; Description of historical church.}
}}
\usage{
historical_churches(coords = F)
}
\arguments{
\item{coords}{show basic data about historical churches in Poznań}
}
\description{
This function download data about historical Churches in Poznan.
}
\details{
Details of usage
}
\examples{
Church <- historical_churches(coords = F)
Church_coord <- historical_churches(coords = T)
}
\keyword{keyword}
