% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_government.R
\name{local_government}
\alias{local_government}
\title{local_government Function}
\format{\describe{
\item{ID}{numeric; ID of District.}
\item{AFDP}{factor; Adaptation for disabled pearson.}
\item{District_no}{numeric; Number of district.}
\item{Residence}{factor; Name of Residence. }
\item{Longitude}{numeric; Longitude of the polls.}
\item{Latitude}{numeric; Latitude of the polls.} 
}}
\usage{
local_government()
}
\description{
This function download data about local government in Poznań.
}
\examples{
Local_Government <- local_government()
}
\keyword{keyword}
