% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_government.R
\name{local_government}
\alias{local_government}
\title{local_government Function}
\format{\describe{
\item{ID}{numeric; ID of District.}
\item{AFDP}{factor; Adaptation for disabled pearson.}
\item{District_no}{numeric; Number of district.}
\item{Residence}{factor; Name of Residence. }
}}
\usage{
local_government(coords = F)
}
\arguments{
\item{coords}{show basic data about local government in Poznań}
}
\description{
This function download data about local government in Poznań.
}
\details{
Details of usage
}
\examples{
Local_Government <- local_government(coords = F)
Local_Government_coords <- Local_Government(coords = T)
}
\keyword{keyword}
