% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monuments.R
\name{monuments}
\alias{monuments}
\title{monuments  Function}
\format{\describe{
\item{ID}{numeric; ID of monument}
\item{Code}{factor; code of monument in Poznan.}
\item{Name}{factor; Name of monument}
\item{Category_builiding}{factor; Category of monument in Poznan.}
\item{City}{factor; City.}
\item{Lang}{factor; Language.}
\item{Address}{factor; Address of monument}
\item{ID_monument}{factor; ID monument in Poznan.}
\item{Description}{factor; Description of monument}
\item{Longitude}{numeric; Longitude of monument.}
\item{Latitude}{numeric; Latitude of monument.}
}}
\usage{
monuments()
}
\description{
This function download data about monuments in Poznan.
}
\examples{
relikt. <- monuments()
}
\keyword{keyword}
