% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_bike.R
\name{paths_bike}
\alias{paths_bike}
\title{paths_bike Function}
\format{\describe{
\item{ID}{numeric; ID of all bike paths in Poznan.}
\item{Name}{character; Name of bike paths.}
\item{Lenght}{character; Lenght of bike paths}
\item{Description}{character; Description of bike paths}
}}
\usage{
paths_bike(coords = F)
}
\arguments{
\item{coords}{show basic data about bike paths in Poznań}
}
\description{
This function download data about bike Paths in Poznan.
}
\details{
Details of usage
}
\examples{
Bike_Paths <- paths_bike(coords = F)
Area_Electoral_coords <- paths_bike(coords = T)
}
\keyword{keyword}
