% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_walking.R
\name{paths_walking}
\alias{paths_walking}
\title{paths_walking  Function}
\usage{
paths_walking(Coord = F)
}
\arguments{
\item{Coord}{show basic_data about trail in Poznań}
}
\description{
This function download data about trail in Poznan.
}
\details{
Details of usage
}
\examples{
Paths <- paths_walking(Coord = F)
Paths_coord <- paths_walking(Coord = T)
}
\keyword{keyword}
