% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticket_sales_points.R
\name{ticket_sales_points}
\alias{ticket_sales_points}
\title{ticket_sales_points  Function}
\format{\describe{
\item{Opening_Hours}{factor; Hours of open.}
\item{Sales_form}{factor; form of sale.}
\item{Ticket}{factor; kind of Ticket.}
\item{Name}{factor; Name.}
\item{Address}{factor; Address of shpo.}
\item{City}{factor;City.}
\item{Lang}{factor; Language.}
\item{PEKKA_Card}{factor; Card.}
\item{ID_Class}{factor; ID of Class.}
\item{Description}{factor; Decsription.}
\item{POS}{factor; Description Class.}
}}
\usage{
ticket_sales_points(coords = F)
}
\arguments{
\item{coords}{show basic data about ticket sales points in Poznań}
}
\description{
This function download data about ticket sales points in Poznan.
}
\details{
Details of usage
}
\examples{
Ticket <- ticket_sales_points(coords = F)
Ticket_coord <- ticket_sales_points(coords = T)
}
\keyword{keyword}
